Refactor search service to comply with new interface
This commit removes the AbstractSearchService and refactors the existing AssetSearchService to comply with the updated SearchServiceInterface. As part of this, it adds a new method 'performSearch' inside the AssetSearchService. Also, this introduces a new SearchProvider and SearchProviderInterface to create different types of searches. An exception has been added for invalid page and page size in PaginatedSearchTrait. Corresponding tests have been added to ensure the accuracy of changes.