Refactor SearchService (#69)
* Remove validation from service.

* Refactor search service to comply with new interface

This commit removes the AbstractSearchService and refactors the existing AssetSearchService to comply with the updated SearchServiceInterface. As part of this, it adds a new method 'performSearch' inside the AssetSearchService. Also, this introduces a new SearchProvider and SearchProviderInterface to create different types of searches. An exception has been added for invalid page and page size in PaginatedSearchTrait. Corresponding tests have been added to ensure the accuracy of changes.

* Apply php-cs-fixer changes

* Update Search.php and AdapterSearchInterface.php return types

* Extract from AdapterSearchInterface to OpenSearchSearchInterface.

* Apply php-cs-fixer changes

* Add Internal.

* Return OpenSearch Interface.

* Fix return type openSearch Interface.

* Apply php-cs-fixer changes

* Style.

* Apply php-cs-fixer changes

* Move source to OpenSearchSearchInterface.

* Remove internal.

* Apply php-cs-fixer changes

---------

Co-authored-by: martineiber <martineiber@users.noreply.github.com>